LinQ and Lambdafied things

+++ Assets/Scripts/Levels/HitObjectLocationManager.cs
@@ -15,10 +15,10 @@ public class HitObjectLocationManager
         _spawnTime = self.SpawnTime;
     }
 
-    public float Distance;
+    public double Distance;
     public Vector3 GetLocation(float audioSourceTime, float speed)
     {
-        Distance = (float)(speed * (-audioSourceTime + _spawnTime) + 5.6 * Assets.Instance.SavedSettings.ElementsSize);
+        Distance = speed * (-audioSourceTime + _spawnTime) + 5.6 * Assets.Instance.SavedSettings.ElementsSize;
         var radians = Rotation * Mathf.Deg2Rad;
         
         var x = (float) (Distance * Math.Sin(radians));
